remove unused code in map_renderer and modify test files
